starting to write testing
